W136 trim FUN_0040c1e0 pointer and local widths

+++ src/Battle/btlFormation.c
@@ -7925,9 +7925,9 @@ u32 func_002c4e50(int param_1,int param_2,u16 *param_3,u32 param_4,short param_5
   char cVar1 = 0;
   u32 uVar2 = 0;
   u32 uVar3 = 0;
-  u32 uVar4 = 0;
+  u16 uVar4 = 0;
   int iVar5 = 0;
-  int lVar6 = 0;
+  long lVar6 = 0;
   u32 uVar7 = 0;
   u32 uVar8 = 0;
   long lVar9 = 0;
