+++ src/Graphics/Model/mdlFile.c
@@ -48252,7 +48252,7 @@ u64 FUN_0034bfc0(u32 param_1)
 
   __int128 auVar2;
 
-  u32 uVar3;
+  u64 uVar3;
 
   u32 uVar4;
